requirements.txt
torch==2.7.0
pillow==11.2.1
accelerate==1.7.0
jsonlines==4.0.0
deepspeed==0.16.9
